backend/requirements.txt
# Core
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.17

# Database
supabase==2.10.0
python-dotenv==1.0.1

# Validation & Serialization
pydantic==2.10.3
pydantic-settings==2.6.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0
httpx>=0.26,<0.28

# Development
ruff==0.8.1
